Support for continued conversations and more
A new major transform called `ConverseLLM` (previously known as
ContinueConversation) that allows to lead continued conversations
with kept history.

This includes extracting reasoning content and other transforms
for working with conversation histories.

`AskPrompt` was kept for backward compatibility. While ConverseLLM is
generally recommended, AskPrompt makes it easy to use the "starts_with"
setting and have it excluded from the output.

Otherwise `AskPrompt` allows to extract the reasoning content to
a selected `reasoning_field` (default strategies use this to save
reasoning content to metadata).

Because it might not be obvious, I added a warning that shows whenever
litellm version is below 1.70.0, as it might cause litellm to not
receive reasoning_content properly. (I'm not 100% sure about which
version is the first to work with gpt-oss).

Additionally, this update includes a json sanitization that prevents
factgenie from crashing when data contains NaN values.

Lastly, there is an extra text processing method that lets you split
a single prompt into many sections (for multi-prompt conversations).